Move cert generator and exporter into the shared package
Phase 1a of carving a reusable host-side cert library out of the UI
extension so a future host CLI (and an eventual non-VS Code distribution
path) can consume the same code. Continues the existing
`@devcontainer-dev-certs/shared` re-export pattern already in use for
loader / pfx / properties / cert-types.

`src/shared/src/cert/exporter.ts` and `src/shared/src/cert/generator.ts`
now hold the canonical implementations; the previous UI-extension copies
become thin re-export shims so existing `./cert/exporter` /
`./cert/generator` imports across the extension and its 18-file test
suite resolve unchanged.

No behavior change: 207 tests, type-check, and full-repo lint all pass.
