+++ src/vscode-ui-extension/src/cert/generator.ts
@@ -1,235 +1,17 @@
-import {
-  AuthorityKeyIdentifierExtension,
-  BasicConstraintsExtension,
-  ExtendedKeyUsage,
-  ExtendedKeyUsageExtension,
-  Extension,
-  KeyUsageFlags,
-  KeyUsagesExtension,
-  SubjectAlternativeNameExtension,
-  SubjectKeyIdentifierExtension,
-  X509CertificateGenerator,
-  cryptoProvider,
-} from "@peculiar/x509";
-import { randomBytes, webcrypto } from "node:crypto";
-import {
-  DevCert,
-  DevKey,
-  ASPNET_HTTPS_OID,
-  CURRENT_CERTIFICATE_VERSION,
-  RSA_KEY_SIZE,
-  SAN_DNS_NAMES,
-  SAN_IP_ADDRESSES,
-} from "@devcontainer-dev-certs/shared";
-
-// `isValidDevCert` and `getCertificateVersion` used to live in this file;
-// they now live in `@devcontainer-dev-certs/shared/cert/validation` so the
-// workspace extension can call the same code path when scanning for a
-// container-side dev cert to push to the host. Re-exported here so existing
-// imports of `./cert/generator` keep working unchanged.
+// Re-export shim: the canonical home for `generateCertificate` is now
+// `@devcontainer-dev-certs/shared`. Keeping this thin re-export so existing
+// `./cert/generator` imports across the UI extension (and its test suite)
+// keep resolving without a sweeping rename. The validation helpers
+// (`isValidDevCert`, `getCertificateVersion`, `computeThumbprint`) used to be
+// defined alongside the generator and were re-exported here for historical
+// reasons — we preserve those re-exports so existing call sites still resolve.
 export {
+  generateCertificate,
   isValidDevCert,
   getCertificateVersion,
   computeThumbprint,
 } from "@devcontainer-dev-certs/shared";
-
-let cryptoProviderConfigured = false;
-function ensureCryptoProvider(): void {
-  if (cryptoProviderConfigured) return;
-  cryptoProvider.set(webcrypto as unknown as Crypto);
-  cryptoProviderConfigured = true;
-}
-
-export interface GeneratedCert {
-  cert: DevCert;
-  key: DevKey;
-  /**
-   * SHA-1 thumbprint, uppercase hex. This is the .NET-compatible
-   * `X509Certificate2.Thumbprint` value used as the X509Store filename
-   * (`{thumbprint}.pfx`). For a stronger cert identifier, use
-   * `cert.thumbprint` (SHA-256).
-   */
-  thumbprint: string;
-}
-
-/**
- * Algorithm choices for `generateCertificate`. Defaults to RSA-2048 to match
- * the historical ASP.NET dev cert format, but the same code path supports
- * ECDSA P-256/P-384/P-521 and Ed25519 for user-managed flows.
- */
-export type GenerateAlgorithm =
-  | { kind: "rsa"; modulusLength?: number }
-  | { kind: "ec"; namedCurve: "P-256" | "P-384" | "P-521" }
-  | { kind: "ed25519" }
-  | { kind: "ed448" };
-
-/**
- * Generate a self-signed certificate matching the ASP.NET Core HTTPS dev
- * cert format (subject, validity, SANs, custom OID, SKI/AKI).
- *
- * The default algorithm is RSA-2048 with SHA-256 signing — a byte-for-byte
- * compatible replacement for the previous `node-forge` path. Pass an
- * `algorithm` to opt into ECDSA / Ed25519 (used by user-managed cert flows
- * that need to round-trip non-RSA keys).
- */
-export async function generateCertificate(
-  notBefore: Date,
-  notAfter: Date,
-  algorithm: GenerateAlgorithm = { kind: "rsa" }
-): Promise<GeneratedCert> {
-  ensureCryptoProvider();
-
-  const { keyPair, signingAlgorithm } = await generateKeyPair(algorithm);
-
-  const serialNumber = generateSerialNumber();
-  const subject = "CN=localhost";
-  const issuer = subject;
-
-  const sanEntries = [
-    ...SAN_DNS_NAMES.map(
-      (dns) => ({ type: "dns" as const, value: dns })
-    ),
-    ...SAN_IP_ADDRESSES.map(
-      (ip) => ({ type: "ip" as const, value: ip })
-    ),
-  ];
-
-  const extensions: Extension[] = [
-    new BasicConstraintsExtension(false, undefined, true),
-    new KeyUsagesExtension(
-      KeyUsageFlags.digitalSignature | KeyUsageFlags.keyEncipherment,
-      true
-    ),
-    new ExtendedKeyUsageExtension([ExtendedKeyUsage.serverAuth], true),
-    new SubjectAlternativeNameExtension(sanEntries, true),
-    new Extension(
-      ASPNET_HTTPS_OID,
-      false,
-      new Uint8Array([CURRENT_CERTIFICATE_VERSION]).buffer
-    ),
-    await SubjectKeyIdentifierExtension.create(keyPair.publicKey),
-    await AuthorityKeyIdentifierExtension.create(keyPair.publicKey),
-  ];
-
-  const cert = await X509CertificateGenerator.create({
-    serialNumber,
-    subject,
-    issuer,
-    notBefore,
-    notAfter,
-    signingAlgorithm,
-    publicKey: keyPair.publicKey,
-    signingKey: keyPair.privateKey,
-    extensions,
-  });
-
-  const devCert = new DevCert(cert);
-  const devKey = await DevKey.fromCryptoKey(keyPair.privateKey);
-
-  return {
-    cert: devCert,
-    key: devKey,
-    thumbprint: devCert.thumbprintSha1,
-  };
-}
-
-async function generateKeyPair(
-  algorithm: GenerateAlgorithm
-): Promise<{
-  keyPair: CryptoKeyPair;
-  signingAlgorithm: Algorithm | RsaHashedKeyGenParams | EcdsaParams;
-}> {
-  const subtle = webcrypto.subtle;
-
-  if (algorithm.kind === "rsa") {
-    const modulusLength = algorithm.modulusLength ?? RSA_KEY_SIZE;
-    const params: RsaHashedKeyGenParams = {
-      name: "RSASSA-PKCS1-v1_5",
-      modulusLength,
-      publicExponent: new Uint8Array([1, 0, 1]),
-      hash: "SHA-256",
-    };
-    const keyPair = (await subtle.generateKey(params, true, [
-      "sign",
-      "verify",
-    ]));
-    return {
-      keyPair,
-      signingAlgorithm: {
-        name: "RSASSA-PKCS1-v1_5",
-        hash: "SHA-256",
-      } as RsaHashedKeyGenParams,
-    };
-  }
-
-  if (algorithm.kind === "ec") {
-    const params: EcKeyGenParams = {
-      name: "ECDSA",
-      namedCurve: algorithm.namedCurve,
-    };
-    const keyPair = (await subtle.generateKey(params, true, [
-      "sign",
-      "verify",
-    ]));
-    return {
-      keyPair,
-      signingAlgorithm: {
-        name: "ECDSA",
-        hash: defaultEcHash(algorithm.namedCurve),
-      } as EcdsaParams,
-    };
-  }
-
-  if (algorithm.kind === "ed25519") {
-    const keyPair = (await subtle.generateKey(
-      "Ed25519",
-      true,
-      ["sign", "verify"]
-    )) as CryptoKeyPair;
-    return {
-      keyPair,
-      signingAlgorithm: { name: "Ed25519" },
-    };
-  }
-
-  if (algorithm.kind === "ed448") {
-    const keyPair = (await subtle.generateKey(
-      { name: "Ed448" },
-      true,
-      ["sign", "verify"]
-    )) as CryptoKeyPair;
-    return {
-      keyPair,
-      signingAlgorithm: { name: "Ed448" },
-    };
-  }
-
-  throw new Error(
-    `Unsupported algorithm: ${(algorithm as { kind: string }).kind}`
-  );
-}
-
-function defaultEcHash(curve: string): string {
-  switch (curve) {
-    case "P-256":
-      return "SHA-256";
-    case "P-384":
-      return "SHA-384";
-    case "P-521":
-      return "SHA-512";
-    default:
-      return "SHA-256";
-  }
-}
-
-function generateSerialNumber(): string {
-  const maxAttempts = 5;
-  for (let attempt = 0; attempt < maxAttempts; attempt++) {
-    const bytes = randomBytes(16);
-    bytes[0] &= 0x7f; // ensure non-negative
-    if (bytes.some((value) => value !== 0)) {
-      return bytes.toString("hex");
-    }
-  }
-  throw new Error("Failed to generate a non-zero certificate serial number.");
-}
+export type {
+  GenerateAlgorithm,
+  GeneratedCert,
+} from "@devcontainer-dev-certs/shared";
